+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## v0.1.23
 
+- Add local data attribute `lecturers`
 - Add support for Campusonline Public REST API, where courses are cached in a database. Which API to use (legacy XML or Public REST) can be
 set using the bundle configuration.
 - Add support for Symfony 7